requirements.txt
matplotlib==3.4.3
numpy==1.21.2
natsort==7.1.1
Pillow==8.3.2
